requirements.txt
diffusers
huggingface_hub
numpy
Pillow
torch
transformers
trl
wandb
peft
